Python/requirements.txt
# requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==3.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
Deprecated==1.2.18
# dnspython==2.8.0 # <-- REMOVIDO
ecdsa==0.19.1
email-validator==2.3.0
# emails==0.6 # <-- REMOVIDO
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
# --- REMOVER LINHAS INVÁLIDAS ---
# pip install aiomysql
# pip install aiosqlite
# --- FIM REMOÇÃO ---
httpx
limits==5.6.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
packaging==25.0
passlib==1.7.4
premailer==3.10.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
# --- NOVAS DEPENDÊNCIAS MFA ---
pyotp==2.9.0 # Biblioteca para gerar/verificar códigos TOTP
qrcode[pil]==7.4.2 # Para gerar URIs de QR Code (instalar com 'pip install qrcode[pil]')
# --- FIM NOVAS DEPENDÊNCIAS ---
requests==2.32.5
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.3